Change int to long

+++ src/main/java/Grammar/Expression/IntLit.java
@@ -5,9 +5,9 @@ import GamePlay.Territory;
 
 import java.util.Map;
 
-public record IntLit(int val) implements Expression {
+public record IntLit(long val) implements Expression {
     @Override
-    public double eval(Map<String, Integer> bindings, Player p, Territory t) {
+    public long eval(Map<String, Integer> bindings, Player p, Territory t) {
         return val;
     }
     @Override
